SSCS-10321 PartyFlagsMapping Refactor
